* 'devel' of https://github.com/arangodb/arangodb:
  fixed shard ordering in Job::clones with consequences for unit tests
  improved debug facilities
  added isLocked() flag for ConditionLocker
  clarify index names
  Redesigned Collection.drop() in RocksDB. Does not write LogData and has a different workflow to create a consistent user-view
  fixed compile with USE_JEMALLOC=OFF
  cppcheck
  updated release notes
  Split tests to fix mac
  typo
  distributeShardsLike is not ignored when prototype does not exist
  fix wrong info. thanks @Simran-B
  fixed issue #2459
  fixed memleak
  Fixed startup exception
  Fixed typedefs
  RocksDB WAL entries for transactions
  Working initial sync
  Fixing incremental sync deletion

+++ lib/SimpleHttpClient/ClientConnection.cpp
@@ -326,15 +326,15 @@ bool ClientConnection::writeClientConnection(void const* buffer, size_t length,
 
 #if defined(__APPLE__)
   // MSG_NOSIGNAL not supported on apple platform
-  int status = TRI_send(_socket, buffer, length, 0);
+  long status = TRI_send(_socket, buffer, length, 0);
 #elif defined(_WIN32)
   // MSG_NOSIGNAL not supported on windows platform
-  int status = TRI_send(_socket, buffer, length, 0);
+  long status = TRI_send(_socket, buffer, length, 0);
 #elif defined(__sun)
   // MSG_NOSIGNAL not supported on solaris platform
-  int status = TRI_send(_socket, buffer, length, 0);
+  long status = TRI_send(_socket, buffer, length, 0);
 #else
-  int status = TRI_send(_socket, buffer, length, MSG_NOSIGNAL);
+  long status = TRI_send(_socket, buffer, length, MSG_NOSIGNAL);
 #endif
 
   if (status < 0) {
